Fixed #169 setting is empty and white
