+++ app/src/main/res/layout/activity_setting.xml
@@ -1,0 +1,151 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:fitsSystemWindows="true">
+  <androidx.constraintlayout.widget.ConstraintLayout
+      android:id="@+id/constraintLayout"
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content">
+
+    <androidx.appcompat.widget.Toolbar
+        android:id="@+id/toolbar"
+        app:titleTextAppearance="@style/ToolbarTitle"
+        style="@style/Toolbar.Common"
+        app:contentInsetStartWithNavigation="0dp"
+        android:theme="@style/Toolbar.Common"
+        app:title="@string/settigs"
+        app:titleMarginStart="80dp"
+        app:layout_scrollFlags="scroll|enterAlways"
+        app:popupTheme="@style/AppToolbarPopup"
+        tools:ignore="MissingConstraints">
+    </androidx.appcompat.widget.Toolbar>
+
+    <androidx.appcompat.widget.AppCompatImageView
+        android:id="@+id/iv_back"
+        android:layout_width="40dp"
+        android:layout_height="40dp"
+        android:layout_gravity="start"
+        android:layout_marginStart="16dp"
+        android:padding="5dp"
+        android:onClick="back"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:srcCompat="@drawable/ic_baseline_keyboard_backspace_24"
+        android:layout_marginLeft="16dp" />
+  </androidx.constraintlayout.widget.ConstraintLayout>
+
+  <LinearLayout
+      android:layout_width="match_parent"
+      android:layout_height="wrap_content"
+      android:orientation="vertical"
+      style="@style/LinearLayout.Base"
+      android:layout_below="@+id/constraintLayout"
+      >
+
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginTop="@dimen/value_20dp"
+        android:orientation="vertical"
+        android:padding="@dimen/value_20dp">
+
+
+
+      <androidx.cardview.widget.CardView
+          android:onClick="notificationAct"
+          android:id="@+id/cv_notification_setting"
+          android:layout_width="match_parent"
+          android:layout_height="wrap_content"
+          app:cardElevation="1dp"
+          app:cardUseCompatPadding="true">
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_margin="@dimen/value_20dp"
+            android:text="@string/notification_settings"
+            android:textColor="@color/black"
+            android:textSize="17sp" />
+
+      </androidx.cardview.widget.CardView>
+
+      <androidx.cardview.widget.CardView
+          android:onClick="about_ac"
+          android:id="@+id/About_us_setting"
+          android:layout_width="match_parent"
+          android:layout_height="wrap_content"
+          app:cardElevation="1dp"
+          app:cardUseCompatPadding="true">
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_margin="@dimen/value_20dp"
+            android:text="@string/about_us"
+            android:textColor="@color/black"
+            android:textSize="17sp" />
+
+      </androidx.cardview.widget.CardView>
+      <androidx.cardview.widget.CardView
+          android:id="@+id/help_setting"
+          android:onClick="helpAc"
+          android:layout_width="match_parent"
+          android:layout_height="wrap_content"
+          app:cardElevation="1dp"
+          app:cardUseCompatPadding="true">
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_margin="@dimen/value_20dp"
+            android:text="@string/help"
+            android:textColor="@color/black"
+            android:textSize="17sp" />
+
+      </androidx.cardview.widget.CardView>
+      <androidx.cardview.widget.CardView
+          android:onClick="share"
+          android:id="@+id/share_setting"
+          android:layout_width="match_parent"
+          android:layout_height="wrap_content"
+          app:cardElevation="1dp"
+          app:cardUseCompatPadding="true">
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_margin="@dimen/value_20dp"
+            android:text="@string/share"
+            android:textColor="@color/black"
+            android:textSize="17sp" />
+
+      </androidx.cardview.widget.CardView>
+
+    </LinearLayout>
+
+  </LinearLayout>
+
+  <Button
+      android:onClick="logout"
+      android:id="@+id/btn_logout"
+      android:layout_width="match_parent"
+      android:layout_height="60dp"
+      android:layout_alignParentBottom="true"
+      android:layout_marginStart="@dimen/default_margin"
+      android:layout_marginLeft="@dimen/default_margin"
+      android:layout_marginEnd="@dimen/default_margin"
+      android:layout_marginRight="@dimen/default_margin"
+      android:layout_marginBottom="@dimen/layout_padding_10dp"
+      android:background="@color/blue"
+      android:text="@string/log_out"
+      android:textColor="@color/white"
+      app:layout_constraintBottom_toBottomOf="parent"
+      app:layout_constraintEnd_toEndOf="parent"
+      app:layout_constraintStart_toStartOf="parent" />
+
+</RelativeLayout>
