--- app/src/main/res/layout/fragment_settings.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".ui.mifos.settings.SettingsFragment">
-
-    <!-- TODO: Update empty settings page -->
-
-</FrameLayout>
