+++ app/src/main/res/menu/menu_nav_drawer.xml
@@ -43,36 +43,11 @@
             android:icon="@drawable/ic_products_black_24dp"
             android:title="@string/products" />
 
-    </group>
-
-    <group android:id="@+id/group2"
-           android:checkableBehavior="single">
-
         <item
             android:id="@+id/item_settings"
             android:checked="false"
             android:icon="@drawable/ic_settings"
             android:title="@string/settings"/>
-        <item
-            android:id="@+id/item_about_us"
-            android:checked="false"
-            android:icon="@drawable/ic_info_black_24dp"
-            android:title="@string/about_us"/>
-        <item
-            android:id="@+id/item_help"
-            android:checked="false"
-            android:icon="@drawable/ic_help_black_24dp"
-            android:title="@string/help"/>
-        <item
-            android:id="@+id/item_share"
-            android:checked="false"
-            android:icon="@drawable/ic_share_black_24dp"
-            android:title="@string/share"/>
-        <item
-            android:id="@+id/item_logout"
-            android:checked="false"
-            android:icon="@drawable/ic_logout"
-            android:title="@string/logout"/>
     </group>
 
 </menu>
