+++ build.gradle
@@ -8,9 +8,11 @@ buildscript {
         mavenCentral()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.3.2'
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
+        classpath 'com.android.tools.build:gradle:4.2.2'
+        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:1.6.10"
         classpath 'com.google.android.gms:oss-licenses-plugin:0.9.5'
+        classpath('com.android.support:multidex:1.0.3')
+
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
     }
